fix(finance): send raw integer nominal from Add/Edit Jenis Pembayaran
The nominal (NOMINAL/JUMLAH) field on the payment-type form auto-formats
as the admin types via CurrencyInputFormatter (e.g. "500000" → "Rp
500.000", Indonesian dot thousand-grouping), and parseCurrency strips
the "Rp " prefix and dots back to the raw value before submit.

parseCurrency returned a double, so the create/update payload carried
500000.0 rather than a clean integer rupiah amount. Cast to int so the
value sent to /payment-types is the raw integer (500000), matching the
field's intent without changing the existing payload shape (Laravel's
numeric rule accepts both, so this is backward compatible).

+++ lib/features/finance/presentation/mixins/payment_form_handlers.dart
@@ -74,9 +74,14 @@ mixin PaymentFormHandlersMixin on ConsumerState<PaymentTypeFormSheet> {
       return;
     }
 
+    // The nominal field auto-formats as the admin types (e.g. "500000" →
+    // "Rp 500.000") via CurrencyInputFormatter. parseCurrency strips the
+    // "Rp " prefix and the dot thousand-separators back to the raw value;
+    // cast to int so the wire payload carries a clean integer rupiah amount
+    // (e.g. 500000) instead of "Rp 500.000" or a 500000.0 float.
     final parsedAmount = CurrencyInputFormatter.parseCurrency(
       amountController.text,
-    );
+    ).toInt();
 
     try {
       final data = <String, dynamic>{
